feat: introduce ftrace marker
Summary:
Adds a class for ftrace marking program flow of KWin. This used in conjunction
with a tool like GPUvis can be very powerful.

I use it at the moment for analyzing the compositing behavior of KWin on X11.
For this I also apply a patch to the XServer adding the same capability [1].

Besides the class also some marking points have been added. Ftracing can be,
if it has been built into the binary, activated via D-Bus.

[1] https://gitlab.freedesktop.org/xorg/xserver/merge_requests/209

Reviewers: #kwin

Subscribers: kwin

Tags: #kwin

Differential Revision: https://phabricator.kde.org/D23114