feat: add wp_viewporter support
With the new interface in Wrapland we can add support for viewports. Here the
specified viewport is being respected by setting the texture coordinates when
rendering with OpenGl and setting the source rectangle with QPainter.

The destination size is already implicitly respected through the size of the
Wrapland surface interface.

Tested with weston-scaler.