feat(xwl): support resolution change emulation
This patch adds functionality to make use of recent XWayland patches for
emulating RandR resolution changes through the Wayland wp_viewporter protocol
extension.

For that a specific window property must be read and the emulated size sent via
X11 protocol to the respective client window.

This patch is WIP because I noticed structural problems that we have with the
current XWayland code. For one KWin is creating one too many parent windows to
work together with the new XWayland code.

Tested with patches to XWayland currently being reviewed and the games
supertux2 and supertuxkart.