feat: delay composite
To minimize latency delay compositing the final picture till shortly before
vblank. This is done by implementing a delay timer with refresh cycle length,
paint duration and a potential vblank gap being factored in.

This is available when buffer swap events are supported, otherwise we fall back
to a simple refresh timer what composites at a random point in time in the
refresh cycle.