feat: implement presentation-time
Provide feedback to clients about the time their content submission was
presented to the user through the presentation-time protocol extension.

The current implementation uses the timestamp of the DRM page-flip event. In
other backends a software timer is used.

In multi-output setups values are still reported back but since we currently
do not support different refresh rates of outputs the implementation must be
revisited once we do.