fix(hw): work around missing mode data in crtc struct
When starting a Wayland session from SDDM the crtc mode data is not complete
for some reason. Maybe this is because of the VT switch. Whatever the reason is
when afterwards we compare the current crtc mode with the modes announced in
the associated connector we can not compare all values. Only compare therefore
a subset which always was available in runtime tests to find the current mode.
Otherwise just fall back to the first one in the list.