ci: setup GitLab CI
Adds a GitLab CI configuration file to run checks on branches and merge
requests. This includes a linter for the commit messages, final build of the
branch and autotests.

In the future additional linters should be introduced, for example cppcheck or
a clang-format based one.

Another goal is at some point to run tests with sanitizers enabled. This is
only a vision though. The current code needs a lot of work for that if it is
possible at all with the many ways in which we depend on Qt.

The 'ci' type is added to the list of allowed commit types.