fix(xwl): handle reset of surface
XWayland clients can be associated with a new surface if the window gets
unmapped and mapped again.

This is in parallel to the Wayland connection communicated via the X11
protocol.

We don't know which event will be first and in my tests it showed that on
certain machiens it was one way or the other.

Therefore remove the assert and handle the unset of the previous surface if
it exists.