ci: handle linter Git remote already existing in cache
The commit message linter directory is cached for subsequent runs. That means
the git remote is already set and trying to set it again yields an error.

Therefore in this case fall back to just setting the url to the current one.
This will succeed if the remote already exists.