build: update version number for branch off
It is not quite clear why there must be two patch version numbers, for example
5.17.80 and 5.17.90 for pre-branch-off and branch-off, respective pre-beta and
beta phase. The beta is already defined by being on the Plasma/5.18 branch.

Some packagers use the 80 number for internal record keeping but it is not per
se necessary.

Let's go with the simpler approach as we do in Wrapland and directly use the 90
patch level from now on.

For the imminent branch-off set the version to 5.17.90. After branch-off on the
master branch we will set it directly to 5.18.90.

In contrast to Wrapland we don't respect SemVer in KWinFT because KWinFT is not
primarily a library (although there are certain parts exported through an API)
but an application.