build: use CMake project 80 patch level in development
Makes it consistent with KDE upstream. We adjust the versions in script anyway,
so it's not much more effort.

Also while CMake does not support SemVer at the moment, if it does at some
point in the future we could just reuse the current setup and write in the
CMake project file something like "5.18.0-0" instead of the current "5.17.80"
to indicate development and then accordingly "5.18.0-beta.0" instead of
"5.17.90" for the beta.