ci: add beta release script
The script must be currently executed locally. It raises CMake project
versions, updates the changelog, tags the commit, branches off a new stable
branch and raises the version again on master for further development.

It uses the last Git tag on master as anchor point to raise the tag version and
CMake versions from and to determine the stable branch name.