refactor(hw): remove output configuration file write
Writing the current configuration in the DRM backend was a workaround to remove
visual artifacts on startup due to asynchronous resizing by KDisplay. But this
increases the overall complexity and the Screens refactor showed that it is
also brittle as making it a noop fixed some instability on output
reconfiguration.