fix: show first item at right position on property list
The first time the list in `Add properties` is shown, the position of the first
delegate is miscalculated as (0, 0) and gets hidden behind the section header.
It only repositions when the delegate is destroyed and re-created, for example,
when setting and deleting a filter text.
(https://bugs.kde.org/show_bug.cgi?id=422289)

This works-around it by delaying the `overlayModel` filtering until the list is
first shown, which makes the delegate position correctly.

BUG: 421583
FIXED-IN: 5.19