refactor(hw): remove X11 EGL backend
The X11 EGL backend is barely used in practice. The switch to activate it was
removed from the UI a long time ago in d52944c2.

To simplify the backend code remove the EGL backend when running as an X11
window manager now completely.

We can move over all methods that were shared with the nested Wayland on X11
backend into this one and by that reduce the inheritance depth by one.