refactor(hw): remove buffer age in windowed X11 backend
The buffer age functionality was only partly setup in the Wayland on X11 nested
backend.

Since this backend is at the moment normally only used for testing remove this
setup code to reduce the code complexity.

If one later finds that the backend is usable also in other contexts than
testing and performance becomes relevant one can add the setup code back and
then add the missing bits.