refactor(hw): remove special Gl strict binding handling
In the EGL backends we had some special handling when strict Gl binding is
active. But the actual impact is unclear. Runtime tests show no difference.

Since we removed the EGL on X11 backend the impact is minimal when removing
this special handling code as it only affects the Wayland on X11 backend.

Remove it therefore now for a reduction of code complexity.