refactor: remove scale method option
This removes the drop-down combobox from the compositing KCM. Instead for
OpenGl we now always try to use smooth scale with Lanczos filter and for our
legacy fallback render backend XRender we use the fast scale method.

The fast scaling on XRender means less pristine graphics in this mode but
since this is only meant as a fallback and not for normal use we live with
that.

Overall this change simplifies our Options and Scene code further.