fix: guard against zero screens
When we query a constrained stacking order we assumed that the screens count is
at least one. This is not true when no screen is connected and we tried to
access a zero-sized vector then.

Guard against this by setting the vector size at least to one also when no
screen is connected.