ci: enforce commit message line length
Until now there was a soft limit of 80 chars on body line length. That does not help much
in practice though because the pre-merge lint will just pass.

Instead now enforce the line length limit. To soften the impact use a 90 char limit for
header, body and footer. This is still less than Angular's current recommendation of 100
chars [1] but works better together with GitLab's UI which otherwise displays horizontal
scroll bars.

This change is analogus to an earlier one in Wrapland.

[1] https://github.com/angular/angular/blob/master/CONTRIBUTING.md#commit-message-format