refactor: adapt Wrapland's new Output master class
Wrapland's Server API has changed in that it offers a central output class that
compositors must use for access to all output-relevant functionality, what
includes the interfaces wl_output, xdg-output and output_device_v1.

This change improves handling and data flow and allows us to delete lots of
code that is now handled internally in Wrapland.