ci: add temporary Wrapland build for windowing project
Build Wrapland from its windowing branch in order to have functional CI
integration for the ongoing windowing project which requires breaking changes
to Wrapland pushed to its parallel windowing branch.

This commit is supposed to be reverted after Wrapland's windowing branch has
been merged to master and when KWinFT's windowing branch is about to be merged.