test: adapt screen edges test
There were some regressions from commits picked from classic KWin. The VDs are
now ordered in a single line instead of in a 2x2 pattern and in one test a
client edge is not reserved anymore.

While the first change can probably easily explained the second one looks to be
more subtle. For now just expect the relevant check to fail here.