feat(effect): add a setting to change the animation curve
This adds a option in the Compositor settings to change the default animation
curve used by KWin effects and scripts.

It only takes place if the effect did not specify a custom curve to use,
therefore it doesn't disrupt effects that specify their own curve, like "Yet
Another Magic Lamp".

Curve defaults to "Linear" as before without the setting.