feat(scene): composite to maximal refresh rate
We may use the maximal refresh rate of all connected outputs on every
refresh to calculate the delay for the next one.

In the future at least on Wayland we want to composite per each
output's individual refresh rate but until then and on X11 in general
the best we can do is aligning our compositing to the maximal refresh
rate of all connected outputs.