refactor: remove X11-only CSD property
The property was added several years ago in f0e1e3187e4 as an aggressive
workaround in combination with a script to suppress GTK's CSD.

The script seems to be gone though and nowadays we are not so aggressive
anymore towards GTK's CSD and on Wayland this property can not have any effect
anyways. So the property and related functions are removed.