refactor: check for popup from win namespace
When determining if a window is a popup do that by calling the generic free
function template win::is_popup.

For that we reroute the call with a forced check on the NET property in any
case and only if by that it is not a popup with a call into the window that
by default in Toplevel returns false but can be overriden by child classes.