refactor: have resize-effect as member
AbstractClient has a boolean variable to indicate if a ongoing resize operation
should be done with the resize effect.

This was in the past a static variable, but we can just as well have it as a
member variable. This ensures there are no side-effects and we can refactor
the code more easily later on.