refactor: split out windowing control class
We add a new class win::control that can be used to store and access common
data relevant for clients that can be controlled directly via the compositor
as in supporting usual desktop functionality like moving, resizing, minimizing
and so on.

The class provides for now a virtual interface due to the Toplevel inheritance.
Setter-functions must be called through the free functions in the win
namespace. Both is not yet the optimal solution and should be improved in the
future, in particular the dynamic inheritance should be avoided at some point.