refactor: set fullscreen in control class
Set the state of a window being fullscreened unified in the control class.
Not all controllable windows can be fullscreened so we reimplement the
function checking this to the control class as a virtual function being
overriden in the child classes. For that we add a new xdg_shell_control
child class.