refactor: remove virtual move alternative
In AbstractClient remove the virtual move function taking x/y-coordinates as
input and only provide the one with QPoint as argument.

There is a certain particularity as the function was overriden by the X11Client
class, not the one with QPoint though. This now changes so that previous calls
to the function taking QPoint are now also handled in X11Client. Some manual
tests showed no problems though.