refactor: remove titlebarPosition function
For at least five years now the function has only returned top-position. In
KDecoration2 the position of the titlebar is handled internally without the
compositor having to care.

It is therefore right to assume that the property is not useful anymore.
Remove the getter and adapt all code using it for branches by directly
running the top-position branch and removing all other ones.