fix: create Deleted control
If the deleted window had a control object reconstruct one in the Deleted too.
This way it can still be accessed in Workspace. We also directly copy the modal
property to it. It might be a good idea in the future to do this with other
properties too instead of having special getters for them in the Deleted.