refactor: remove desktops list
In order to simplify the Workspace code remove the separate desktops list
that contained clients acting as a desktop, that is in a Plasma session
plasmashell.

This list was only used on X11 though. To align behavior on X11 and Wayland
just loop through all clients. In the future it might make sense to restrict
this again to X11 clients only though, or if possible to only do it when
running as an XWM only.