refactor: remove X11 clients list
Do not store the X11 clients in a separate list. Only store them in the list
for all clients.

This is a preparational step for later structural changes to Workspace and
unifies the code. But for now this means sometimes we still have to cast to
X11Client when querying the list for all clients.