refactor: use transient class in remnants
Unify code paths by making use of the transient interface in remants just like
with other windows.

Transient relations are simply ported over. This ways emitting the
transientChanged signal can be moved into the transient class too and we have a
clear flow from adding a window as child to removing it again at last at the
destruction of the remnant.