refactor: remove WaylandServer init function
It is considered bad practice to use init functions. Move the initialization
to the constructor and throw an exception in case the server can not be
started.

Unit tests had to be adapted to this new structure. The socket name now is
implicitly set through the name of the test what reduces the boiler plate
code and risk of regressions. Additionally we use a template function to
reduce the size of the test macro that gets expanded into all of the tests.