fix(scene): paint subsurface quads
With subsurfaces being annexed transient children we go through the stacking
order when painting with OpenGl, remove all these from the list and instead
paint them as part of their parent surface.

This paint process is done by creating quads for all subsurfaces and adding
them to the list of the parent.