refactor: replace activation enum with booleans
The enum was complicated as it used a flag that defined through another one.
The enum was used only for a single function call. Replace it just with two
booleans what simplifies the code and fixes an issue with plasmoids that was
introduced on refactoring because the enum value that was logically or-ed
was tested the wrong way.