refactor: remove client content position getter
The getter was only used when constraining the pointer input. In most cases the
call was equivalent with getting the negative client position. When we deal
with an Xwayland window it is unclear if that really makes a difference.

For simplification replace its usage with just a call to the client position.