refactor: remove support for _NET_WM_FRAME_OVERLAP
Support for this property was implemented over 10 years ago in 0a8c06b0 and the
motivation for that was better support of clients written against Windows as
similar functionality is available on that (DwmExtendFrameIntoClientArea).

There is a proposal to include the hint into the official EWMH spec [1] but
looking at the current spec [2] this was unsuccesful.

Since it did not became a standard and I could not find any actual usage in any
clients it is suspected this property is not used anymore or was never actively
used. Also it is X11-only. Following that we remove its implementation.

[1] https://mail.gnome.org/archives/wm-spec-list/2009-November/msg00009.html
[2] https://specifications.freedesktop.org/wm-spec/latest/