refactor: overhaul windows geometries
This is a grand overhaul of the internal geometry handling of windows in
particular under X11.

Window geometries go through an up-to three-stage transformation process:
1. Update of the geometry by user request or internal logic.
2. Geometry pending sync from the client.
3. Geometry being applied to our rendering.

The handling of geometries is unified between all window types by being
processed through these three stages. For that we have a single entry point for
a geometry change which is the setFrameGeometry function.

In all three stages the different types of geometries (frame, client, render)
are calculated from a single source of truth, which is usual the cached frame
geometry. By that we improve cache coherence and geometry updates can be viewed
as atomic. Additionally we reduce the amount of virtual functions we require.
In particular the virtual bufferGeometry function has been removed.

X11 Sync requests are now sent out without fallback timer. It is expected
clients will reply with the latest received one on Wayland while on X11 we
always do a round-trip on resize operations before sending a new sync in order
to set the correct server geometry. Further on X11 we limit our resize
configure notifies with a timer in order to not flood the client.

Updating the server geometry on X11 has been simplified. In particular it is
ensured that windows with GTK frame extents are handled correctly and in a
similar way as we do on Wayland.