test: repeat screenlock unlock request
KScreenLocker does not handle unlock requests via logind reliable as it sends
a SIGTERM signal to the lock process which sometimes under high system load is
not received and handled by the process.

It is unclear why the signal is never received but we can repeat sending the
signal mutliple times assuming that after few tries one of them will be
received.