feat: apply Initially as default value of new Rules Properties
Change the default value from "Do Not Affect" to "Apply Initially" when
adding new properties in Window Rules.

REASON: prior "Do Not Affect" as default and occasionally I forget to
change the setting to something else when adding new properties.
Unfortunately, this means I forget to do it and so it isn't saved when
I close the window. I will then go back in to change it to Apply Initially
or in some cases have to add the new property again as it wasn't saved in
the set.

I think if Apply Initially were to be the default it would be an easier
and faster experience to add new properties.