feat: add Wayland wrapper binary
Plasma session now calls into a wrapper binary that we need to provide.
This wrapper creates a Wayland socket on its own and then forks the compositor.

The purpose is to recover from compositor crashes. Usefulness of this aside in
any case we need to provide the wrapper binary since Plasma expects it.