test: init Xwayland selections test
The initialization of the test was skipped as the message stated due to
failures on KDE's CI "due to unknown reasons". With recent change to the
socket creation this has become necessary though to make the test not
assert on a non-existing display object in Wrapland.

It is assumed the issue with CI has been solved already earlier and skipping
the initialization phase can therefore simply be omitted.