refactor: split up size for client size function
Improve the internal logic by splitting the function up into three separate
functions where the two helper functions are being called from the main
function if required.

This allows us to remove some local variables to reset the position just in
case and instead just do not set it in the first place.