refactor: remove virtual client size function
Instead convert from frame to client size when required. This reduces
the code complexity by removing a virtual function an ensures single source of
truth to the size data.

Note that the special case for a window being a remnant is removed by this
change what might lead to a regression in this case as in the content the
client provides being rendered stretched for a wrong size.