refactor: remove client pos function
Instead convert from frame to client position when required. This reduces
the code complexity and unifies our interface.

Note that the special case for a window being a remnant is removed by this
change what might lead to a regression in this case as in the content might be
rendered with an offset.