fix: limit syncless resize updates
When a client does not support synced resizes on X11 limit the resize rate
such that clients won't be flooded with geometry changes.

That might lead to problems in clients not handling such well. The only example
client where this could be observed though was the Steam client which did only
update its child windows with a huge delay after a resize operation.