fix: determine indirect transient followers
The has_child function had at some point in the past an override to find
transients X11 windows that are indirect children e.g as a child of a child
depending on the value of a boolean trap.

This overload was lost at some point which ultimately means that the code path
regressed. For example Kicker submenues were as second-level children and by
virtue of being tool-windows hidden again through update_tool_windows(..).

This patch replaces the calls to has_child with a call to is_follower_of, which
not only makes the overload unnecessary but is in general a better API.