fix(space): send only children with control to other screen
A transient child window might be without control when it is an xdg-popup or a
subsurface, in general an annexed child.

Check that control exists when looping over the children, otherwise omit
sending the child to the other screen. The geometry of such annexed children
will be changed implicitly through their parent-child relation.