feat: remove shading windows
To shade a window, that is hide its window content and only show the top bar of
its decoration is a long-standing capability of window managers on X11. It is
also part of EWMH [1].

Despite this the functionality is not commonly used or even known by many users
and with CSD it is not possible in general to the point that we still don't
even have it implemented for windows with SSD on Wayland.

Additionally the implementation we had is was very invasive increasing the code
complexity tremendously.

[1] https://specifications.freedesktop.org/wm-spec/wm-spec-1.3.html#idm45841377105904