fix: configure on no-change maximize requests
When a client requests to maximize or unmaximize a window the compositor is
supposed to send a configure event even if no change occured.

This change ensures a configure event is sent but in case the first sent size
was empty we might send on second event a non-empty size when the client
committed in-between with a buffer attached. This needs to be fixed later on.

Another question is if we should handle the fullscreen request in the same
manner. The spec implies this to a certain degree but is not as explicit about
it as with maximization requests.