fix: add fallback timer for windows not syncing back initially
Apps built with Electron 9 do not reply to the initial _NET_WM_SYNC_REQUEST
request. This could be confirmed with Discord and balenaEtcher.

For such apps add a fallback timer on first sync to force the window to be
ready for painting after a short duration. This way we can still sync the first
appearance of other apps correctly but force it for Electron 9 apps. Afterwards
we continue with all apps in the same way.

It is assumed this workaround can be removed once the most common Electron apps
are based on a more recent version of Electron.