fix: always send a final synthetic configure event
On a maximization or fullscreen change borders might change. In this case we
must update the server geometry. This seems to not always trigger a full
geometry update at the moment, as in the geometry might not be shifted to
adapt to the changed borders.

Some clients for example IntelliJ IDEA also seem to expect a final configure
event for a full update.

To ensure in both cases the client has received the updated geometry for now
always send a final synthetic configure event in the end of a setFrameGeometry
call.

Sending such an event should always be ok but in the future we might still want
to send it more specifically, i.e. only when really required, so our internal
implementation is restrained.