fix: unset pseudo-max restore geometries on taking control
An X11 client might only provide its maximized geometry on mapping. In this
case unset it so we do a restore to the default size and placement later on.

An X11 client that showed this behavior was the Element desktop app when
minimizing it from its maximized state directly to tray and then opening it
again from there.