build: use a shared integration test library
We compiled into every autotest a static version of the library. This was
always taking a lot of disk space but with the recent change to link several
internal plugins statically the size increased even more.

Instead of static linking make the integration test library a shared
library which is not installed but can still be linked from the integraiton
tests in the build directory. To allow us to do that some additional entities
needed to be exported.

Here the build directory size decreased from 2.7G to 1.3G.