fix(input): connect directly to deco geometry changes
In 507d83fb a connection was added to update deocration when the geometry of
the window changes. That connection was explicitly made queued. It is not clear
why it was done this way but in the context it seems unnecessary.

Also there was a crash reported due to the window being invalid at the time the
signal handler was called in the event loop. In theory that shouldn't happen as
a connection is capped once the window is destroyed, but making it an direct
connection I'm more sure this is actually the case.