fix: prolong sync fallback time
For Electron 9 apps a timer was added in 851f3be8 to force showing them after
33 ms as they don't react to the first sync request. The 33 ms are too short
on low-speced systems leading to artifacts with other apps as their shadow
might be shown before their content is ready.

As the timer is only useful for few apps and the delay is only relevant on
mapping, prolong the fallback time to 1000 ms. That should other apps give
more than enough time to react to the first sync request and show their content
together with the shadow at the same time.