fix: omit trying to minimize transient children without control
When minimizing or unminimizing a window we have to do this for all transient
children too that are not annexed.

A transient child might be not annexed and still without control though, for
example a deleted one. This will lead to a crash.

Instead of omitting annexed children only, omit all windows without control,
what includes annexed children.