ci: inhibit pages job on Coverity Scan
The Coverity Scan nightly build has been failing silently since the
introduction of coverage reporting. Some tests showed that this is likely due
to the pages job not being inhibited explicitly on the nightly build.

Do this now and use this opportunity to also inhibit the clang-format job on
this build as it is not needed to run there.