feat: add wlroots display backend
Provides a display backend based on wlroots. Allocation of buffers is still
done by us and then handed over to wlroots via dmabuf.

The backend is used by default on DRM and in nested sessions together with our
wlroots input backend code. To disable it set `KWIN_WLROOTS_BACKEND` to 0 for
using neither the wlroots display nor input backends or to `i` to only use the
input backend. Using the input backend only is as before only possible on DRM.