feat: add wlroots input backend
Adds structure for making use of wlroots as input source. At the moment this
only works with libinput, i.e. a DRM session, as in other cases like nested
operations the input backend is combined with the display backend.

The wlroots input backend is by default used in a DRM session, but can be
disabled by setting the envrionment variable `KWIN_WLROOTS_BACKEND` to 0.