test: remove QPainter scene tests
With the wlroots backend we do not provide QPainter rendering anymore. To use
the backend with our integration tests only tests that can run under OpenGl are
supported.

Therefore remove the specific QPainter tests. Corresponding versions for OpenGl
were already available. So we do not lose anything substantial as the QPainter
rendering mode is supposed to be decommissioned later on.