perf(scene): schedule repaints only on relevant outputs
When a window requires scheduling a repaint only do this on relevant outputs,
that are outputs the window geometry intersects with.

This is only possible on Wayland with per-output rendering. We introduce a new
function in Composite class taking the window requesting the repaint as an
argument. The Wayland child class overrides this.