fix(hw): omit buffer age usage when hardware does not transform
When our hardware does not transform, we have to render to the intermediate
frame buffer. In this case it has shown that usage of buffer age leads to
artifacts with some effects like wobbly windows.

As no suitable solution has been found yet for now deactivate the usage of the
buffer age extension in this case. That induces a performance penalty. A TODO
has been added so we revisit this issue later on.