feat(input): add input control
In our new input platform code provide infrastructure to query characteristics
of input devices and control their behavior.

This is primarily aimed at the wlroots backend with libinput at the moment.
The new infrastructure is adaptive enough though to handle other backends in
the future though.

In contrast to the old code in the libinput backend this is a huge step forward
as not only is the backend code decoupled but also any intergration with D-Bus
is  and can be added afterwards interfacing only with the generic control API.