feat(input): add switch devices
Switch devices are required for certain advanced actions. Add them as a new
class of devices to the input backend. They are represented in wlroots by
separate structs too. So this maps well. For our DBus interface we add the
necessary infrastructure too and can now correctly hook up the status getters.