feat(input): provide D-Bus control interface
For our new input platform code add a D-Bus interface to control the
characteristics of input devices.

This new interface is backwards-compatible with the previous interface, which
was coupled with the libinput code.

The new interface implementation is decoupled from the actual backend in use.
The multiplexing of different device classes to the single class of devices on
D-Bus is done directly through pointers in the dbus::device class.