fix: create backend session only once
In the wlroots backend we must ensure a wlr_session is only created once. When
the rendering backend is DRM, this backend already created one. So we must set
this as the native backend for our session, so that we don't create a second
one on the call to takeControl.