fix: set dummy EGL surface
The abstract EGL backend expects an EGL surface to be set for rendering. That
is a legacy assumption from X11 where we did not render per output.

Instead of setting this to one of the outputs' surfaces we set it to a dummy
EGL urface which we create at the beginning of our runtime and keep around
until the end. This is simpler than trying to set and unset on output creation
or destruction and also allows to startup without output connected.