feat(hw): remove legacy fbdev backend plugin
We want to use in the future the wlroots backend exclusively for our Wayland
session. While wlroots does not support running on fbdev we remove our legacy
plugin anyway for consistency and downsizing the code base.

Running on fbdev is normally not really relevant ant it is assumed that it will
become even less important in the future as remaining use cases, like running
Linux with nomodeset [1], are being ported to DRM.

[1] https://patchwork.freedesktop.org/series/78817/