refactor: compile X11 backend plugin as render backend
We compile the X11 backend plugin directly into our kwin_x11 binary and
respective tests in form of a static library and do not load it as a
plugin dynamically at runtime anymore.

This allows us to remove all remaining backend plugin infrastructure. We can
also simplify platform destruction, as this is now done on Wayland and X11 in
the same way: the application instance owns it.