fix(input): set output for touch devices
To send touch events relatively to the output in logical space the output must
be determined. For that in the new input platform try to determine it similarly
like in the old libinput code first by name, and if not available by additional
logic comparing the size and if the system has an internal screen.