refactor: add wlr_backend getter
Add a function that returns a wlr_backend depending on a check function, that
the caller must provide. If no backend is found null is returned.

For multi-backends, the first individual backend the check succeeds for, is
returned.

We use it directly for the headless backend check. Another wrapper call for
checking for a DRM backend is also provided.