refactor(xwl): restructure xfixes notify handling flow
Instead of setting an abstract Selection to either clipboard or a drag-and-drop
operation simply return a call to the handle function directly. If the object
does not exist, return in the function template early.

This will allow us to make the virtual function doHandleXfixesNotify a normal
member function of the child class.