refactor(xwl): handle selection dependent X11 messages in function templates
Some handling of X11 messages depends on the selection, as clipboard and dnd-
operations react differently.

Still we want to move this handling out into free function templates to later
share clipboard selection functionality with the primary selection.

To facilitate that we employ template specializations of the functions for the
Dnd class.