refactor(xwl): handle X11 offer changes in function template
Handling X11 offers depends on the selection, as clipboard and dnd-operations
react differently.

Still we want to move this handling out into a free function template to later
share clipboard selection functionality with the primary selection.

To facilitate that we employ a template specialization of the function for the
Dnd class.

For the clipboard we require to make the calls into Wrapland::Server::Seat
selection setter and the WaylandServer internal device managers variable to
later add the primary selection. Use getters for that in the class.