refactor(scene): store Compositor pointer in Application
Our goal in the end is to make the Compositor a normal member of some class,
that gets created on startup and its lifetime managed by that class or is
self-managed.

For now move the pointer to the Compositor simply to the Application class as
this is guaranteed to exist. Later on we might relocated it and manage its
lifetime appropriately.