style: rename function to continue startup on Wayland
After the scene is created we signal back in order to create the workspace.
This function is therefore misnamed. We might create Xwayland integration
in between, but it is also ok to just name the workspace creation in the
function name as this is the main thing happening in the end.