refactor: split out X11 event filter from Workspace
Instead of managing all kind of event filters in the Workspace class split them
out into a separate event filter manager class which then is not owned by the
Workspace anymore but by the Application itself.

This will allow us to create and destroy event filters independent of the
creation of a Workspace instance.