refactor: load options early in Application
The options values were reparsed in the Workspace ctor in a different thread
and then loaded. It is assumed the threading is premature optimization.
Additionally reparsing the configuration is unnecessary as we only create the
Workspace class once in the beginning.

Instead load the options earlier directly after creating the options object in
the Application class as not only the Workspace class is dependent on these
values.