refactor: always create compositor before workspace
The startup sequence is changed to align it in the X11 session with how we do
it in the Wayland session. We create the Compositor before the Workspace
and after the Workspace has been created too then start the compositing.

In our integration tests we have to handle the cases without and with Xwayland
being started now differently, as without it we go through the startup without
any delay, so the workspace creation spy has already fired.