refactor: create screen locker watcher earlier
Move the creation of the ScreenLockerWatcher to the start function. This
function is usually called before createInput is called but all necessary
dependencies for the screen locker watcher are already created, in particular
the Wayland server instance.

The motivation is to thin out the createInput function over time until it can
be removed.