refactor: remove static cursor variable
In general we want to move away from singletons. The cursor instance has been
already restructured to be owned by the input platform. Now the static variable
is removed and calls to it are replaced with a single static getter retrieving
the cursor from the input platform via the Application instance.