refactor(input): let Wayland cursor own the cursor image
The cursor image was owned by the pointer redirect. That is not intuitive.
Instead let the input::wayland::cursor child of input::cursor own it.

We must ensure that on startup all data is available at the right point in
time. That means also to wait on the Workspace being created, so the
connections can be setup.