fix: handle exception of Wayland server constructor
On Wayland server creation an exception might be thrown. This exception should
be handled for compliance. Coverity Scan raised a warning about the unhandled
exception.

In case the exception is raised we just exit in the integration tests. And a
critical warning is printed out from the Wayland server function which raises
the exception.